refactor(server): move python api out
